New Dinning Page + Button GUI Changes
Keenan Changed the Button CSS so they are no longer default buttons.

I setup a new Gallery Format for the Dinning Page.

We still need at least one .ASP Classic Page/Utility

And preferably a working Tour scheduling//Utility that allows the user
to send an email to the company that tells us:
What day:
What Time:
# of people:
And based on # of People the price should be replied to them upon
submitting//preview.

We can work together on that part.